Develop

+++ database/migrations/2019_10_18_054933_create_bills_table.php
@@ -14,13 +14,13 @@ class CreateBillsTable extends Migration
     public function up()
     {
         Schema::create('bills', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->double('total_price', 15, 2);
             $table->double('total_payment', 15, 2);
             $table->integer('discount');
             $table->string('note', 200)->nullable();
 
-            $table->unsignedInteger('order_id');
+            $table->unsignedBigInteger('order_id');
             $table->foreign('order_id')
                 ->references('id')
                 ->on('orders')
